Changes Made:
Added PID algorithm to huskylens program.  We set up the PID variables, and initiallized the motors: rightFront, rightBack, leftBack, leftFront.  We modified the program to where if an object (Id1) was recognized, the motors would turn the robot, making the object xvalue = 160 (the center of the view).

Problems:

Robot overshot and lost object Id1 (meaning it kept spinning, looking for the object, which it never did). Potential problem may be speed.